backend/requirements.txt
flask==2.2.5
flask-cors==3.0.10
torch==2.2.0
torchvision==0.17.2
Pillow==10.2.0
numpy==1.26.4

opencv-python==4.5.3.56
scikit-image==0.18.3 